requirements.txt
PySide6>=6.5.0

# Real-ESRGAN Dependencies (für inference_realesrgan_video.py)
# PyTorch (kompatible Versionen)
torch>=1.7
torchvision>=0.8.1
# Real-ESRGAN spezifische Packages
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.5
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.3.0
# ffmpeg-python (optional, für Video-Verarbeitung)
ffmpeg-python>=0.2.0

# Stable Diffusion für Cover-Generierung
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0

# Build-Tools
PyInstaller>=5.13.0

